telegram-bot-max/requirements.txt
python-telegram-bot==21.10
firebase-admin==6.6.0
httpx==0.28.1
python-dotenv==1.1.0
anthropic==0.52.0
openai==1.82.0
fastapi==0.115.0
uvicorn==0.34.0
python-multipart==0.0.20
